Drop support for closures.
