+++ machine/Cargo.toml
@@ -10,6 +10,7 @@ repository = "https://github.com/ipfs-rust/netsim-embed"
 [dependencies]
 async-global-executor = "2.0.2"
 async-io = "1.3.1"
+async-process = "1.1.0"
 futures = "0.3.13"
 ioctl-sys = "0.7.1"
 libc = "0.2.90"
