+++ src/lib.rs
@@ -1,6 +1,5 @@
 use async_process::Command;
 use futures::channel::mpsc;
-use futures::io::BufReader;
 use futures::prelude::*;
 use netsim_embed_core::*;
 pub use netsim_embed_core::{Ipv4Range, Wire};
@@ -9,9 +8,9 @@ use netsim_embed_machine::*;
 use netsim_embed_nat::*;
 use netsim_embed_router::*;
 pub use pnet_packet::*;
-use std::io::Write;
+use std::fmt::Display;
 use std::net::Ipv4Addr;
-use std::process::Stdio;
+use std::str::FromStr;
 
 pub fn run<F>(f: F)
 where
@@ -106,7 +105,12 @@ pub struct NetworkBuilder<C, E> {
     networks: Vec<Network<C, E>>,
 }
 
-impl<C: Send + 'static, E: Send + 'static> NetworkBuilder<C, E> {
+impl<C, E> NetworkBuilder<C, E>
+where
+    C: Display + Send + 'static,
+    E: FromStr + Send + 'static,
+    E::Err: std::fmt::Debug + Send,
+{
     pub fn new(range: Ipv4Range) -> Self {
         let router = Ipv4Router::new(range.gateway_addr());
         Self {
@@ -117,21 +121,13 @@ impl<C: Send + 'static, E: Send + 'static> NetworkBuilder<C, E> {
         }
     }
 
-    pub fn spawn_machine<B, F>(&mut self, config: Wire, builder: B) -> Ipv4Addr
-    where
-        B: FnOnce(mpsc::UnboundedReceiver<C>, mpsc::UnboundedSender<E>) -> F + Send + 'static,
-        F: Future<Output = ()> + Send + 'static,
-    {
+    pub fn spawn_machine(&mut self, config: Wire, command: Command) -> Ipv4Addr {
         let (iface_a, iface_b) = config.spawn();
         let (cmd_tx, cmd_rx) = mpsc::unbounded();
         let (event_tx, event_rx) = mpsc::unbounded();
         let addr = self.range.random_client_addr();
         let mask = self.range.netmask_prefix_length();
-        async_global_executor::spawn(async_global_executor::spawn_blocking(move || {
-            let join = machine(addr, mask, iface_b, builder(cmd_rx, event_tx));
-            join.join().unwrap();
-        }))
-        .detach();
+        let _ = machine(addr, mask, iface_b, command, cmd_rx, event_tx);
         let machine = Machine {
             addr,
             tx: cmd_tx,
@@ -140,50 +136,6 @@ impl<C: Send + 'static, E: Send + 'static> NetworkBuilder<C, E> {
         self.machines.push(machine);
         self.router.add_connection(iface_a, vec![addr.into()]);
         addr
-    }
-
-    pub fn spawn_machine_with_command(&mut self, config: Wire, mut command: Command) -> Ipv4Addr
-    where
-        C: std::fmt::Display,
-        E: std::str::FromStr,
-        E::Err: std::fmt::Debug + Send,
-    {
-        self.spawn_machine(
-            config,
-            |mut cmd: mpsc::UnboundedReceiver<C>, event: mpsc::UnboundedSender<E>| async move {
-                command.stdin(Stdio::piped()).stdout(Stdio::piped());
-                let mut child = command.spawn().unwrap();
-                let mut stdout = BufReader::new(child.stdout.take().unwrap()).lines().fuse();
-                let mut stdin = child.stdin.unwrap();
-                let mut buf = Vec::with_capacity(4096);
-                loop {
-                    futures::select! {
-                        cmd = cmd.next() => {
-                            if let Some(cmd) = cmd {
-                                buf.clear();
-                                tracing::trace!("{}", cmd);
-                                writeln!(buf, "{}", cmd).unwrap();
-                                stdin.write_all(&buf).await.unwrap();
-                            } else {
-                                break;
-                            }
-                        }
-                        ev = stdout.next() => {
-                            if let Some(ev) = ev {
-                                let ev = ev.unwrap();
-                                if ev.starts_with('<') {
-                                    event.unbounded_send(ev.parse().unwrap()).unwrap();
-                                } else {
-                                    println!("{}", ev);
-                                }
-                            } else {
-                                break;
-                            }
-                        }
-                    }
-                }
-            },
-        )
     }
 
     pub fn spawn_network(&mut self, config: Option<NatConfig>, mut builder: NetworkBuilder<C, E>) {
